src/requirements.txt
Django==1.7.1
MySQL-python==1.2.5
Twisted==15.0.0
argparse==1.2.1
django-cors-headers==0.13
djangorestframework==2.4.4
ecdsa==0.13
lxml==3.4.1
paramiko==1.15.2
pexpect==3.3
pycrypto==2.6.1
pylibmc==1.4.1
requests==2.4.3
wsgiref==0.1.2
zope.interface==4.1.2
